[REFACTOR] move Freeze to common location
