+++ pandora-client-web/src/routing/Routes.tsx
@@ -19,6 +19,7 @@ import { AccountProfileScreenRouter, CharacterProfileScreenRouter } from '../com
 import { ModalDialog } from '../components/dialog/dialog';
 import { useObservable } from '../observable';
 import { SpaceJoin } from '../ui/screens/spaceJoin/spaceJoin';
+import { Freeze } from '../ui/components/common/freeze';
 
 // Lazily loaded screens
 const Management = lazy(() => import('../components/management'));
@@ -198,31 +199,6 @@ function DeveloperRoutes(): ReactElement {
 	return (
 		<Suspense fallback={ <div>Loading...</div> }>
 			<Management />
-		</Suspense>
-	);
-}
-
-const infinite = new Promise(() => { /** */ });
-
-function Suspender({ freeze, children }: {
-	freeze: boolean;
-	children: React.ReactNode;
-}) {
-	if (freeze) {
-		// eslint-disable-next-line @typescript-eslint/no-throw-literal
-		throw infinite;
-	}
-	return <>{ children }</>;
-}
-
-function Freeze({ freeze, children, placeholder = null }: {
-	freeze: boolean;
-	children: React.ReactNode;
-	placeholder?: React.ReactNode;
-}) {
-	return (
-		<Suspense fallback={ placeholder }>
-			<Suspender freeze={ freeze }>{ children }</Suspender>
 		</Suspense>
 	);
 }
