+++ pandora-client-web/src/ui/components/common/freeze.tsx
@@ -1,0 +1,26 @@
+import React, { Suspense } from 'react';
+
+const infinite = { then: () => { /** noop */ } };
+
+function Suspender({ freeze, children }: {
+	freeze: boolean;
+	children: React.ReactNode;
+}) {
+	if (freeze) {
+		// eslint-disable-next-line @typescript-eslint/no-throw-literal
+		throw infinite;
+	}
+	return <>{ children }</>;
+}
+
+export function Freeze({ freeze, children, placeholder = null }: {
+	freeze: boolean;
+	children: React.ReactNode;
+	placeholder?: React.ReactNode;
+}) {
+	return (
+		<Suspense fallback={ placeholder }>
+			<Suspender freeze={ freeze }>{ children }</Suspender>
+		</Suspense>
+	);
+}
